change mpc listall handling
This change improves the speed of this script. Without any arguments,
mpc listall prints all music files starting from the root of the music
directory. In my case, that includes .git, which is used by git-annex.
This change removes the need to filter out all the lines containing
".git" after running $list_songs_cmd because they are never added to the
list to filter down in the first place.