add org-crypt-diff.pl
slow diff of org files with crypt blocks

This works, but it can be pretty slow if a given org file has more than
1 org-crypt block. You can speed this up by using enabling git
attribute's cachetextconv variable, but this sort of defeats the purpose
of encryption. It's probably best to disable this most of the time, but
enable when I want to run something like git log -p on some set of org
files containing crypt blocks.